travel_planner2/pyproject.toml
[project]
name = "travel_planner2"
version = "0.1.0"
description = "travel-planner2 using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.186.1,<1.0.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "googlemaps>=4.10.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0"
]

[project.scripts]
travel_planner2 = "travel_planner2.main:run"
run_crew = "travel_planner2.main:run"
api = "travel_planner2.main:run_api"
train = "travel_planner2.main:train"
replay = "travel_planner2.main:replay"
test = "travel_planner2.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
